+++ src/compiler/codegen_c_json_parser.c
@@ -12,7 +12,7 @@
 #define PRINTLN_SPMAX 64
 static char println_spaces[PRINTLN_SPMAX];
 
-static void println(output_t *out, const char * format, ...)
+static void println(fb_output_t *out, const char * format, ...)
 {
     int i = out->indent * out->opts->cgen_spacing;
     va_list ap;
@@ -46,7 +46,7 @@ static void println(output_t *out, const char * format, ...)
  * in ignored fields.
  */
 
-static int gen_json_parser_pretext(output_t *out)
+static int gen_json_parser_pretext(fb_output_t *out)
 {
     println(out, "#ifndef %s_JSON_PARSER_H", out->S->basenameup);
     println(out, "#define %s_JSON_PARSER_H", out->S->basenameup);
@@ -60,7 +60,7 @@ static int gen_json_parser_pretext(output_t *out)
     return 0;
 }
 
-static int gen_json_parser_footer(output_t *out)
+static int gen_json_parser_footer(fb_output_t *out)
 {
     gen_pragma_pop(out);
     println(out, "#endif /* %s_JSON_PARSER_H */", out->S->basenameup);
@@ -360,7 +360,7 @@ static void clear_dict(dict_entry_t *dict)
     }
 }
 
-static int gen_field_match_handler(output_t *out, fb_compound_type_t *ct, void *data, int is_union_type)
+static int gen_field_match_handler(fb_output_t *out, fb_compound_type_t *ct, void *data, int is_union_type)
 {
     fb_member_t *member = data;
     fb_scoped_name_t snref;
@@ -611,7 +611,7 @@ repeat_nested:
     return 0;
 }
 
-static void gen_field_match(output_t *out, fb_compound_type_t *ct, void *data, int hint, int n)
+static void gen_field_match(fb_output_t *out, fb_compound_type_t *ct, void *data, int hint, int n)
 {
     println(out, "buf = flatcc_json_parser_match_symbol(ctx, (mark = buf), end, %d);", n);
     println(out, "if (mark != buf) {"); indent();
@@ -620,7 +620,7 @@ static void gen_field_match(output_t *out, fb_compound_type_t *ct, void *data, i
 }
 
 /* This also handles union type enumerations. */
-static void gen_enum_match_handler(output_t *out, fb_compound_type_t *ct, void *data, int unused_hint)
+static void gen_enum_match_handler(fb_output_t *out, fb_compound_type_t *ct, void *data, int unused_hint)
 {
     fb_member_t *member = data;
 
@@ -655,7 +655,7 @@ static void gen_enum_match_handler(output_t *out, fb_compound_type_t *ct, void *
     }
 }
 
-static void gen_enum_match(output_t *out, fb_compound_type_t *ct, void *data, int hint, int n)
+static void gen_enum_match(fb_output_t *out, fb_compound_type_t *ct, void *data, int hint, int n)
 {
     println(out, "buf = flatcc_json_parser_match_constant(ctx, (mark = buf), end, %d, aggregate);", n);
     println(out, "if (buf != mark) {"); indent();
@@ -663,7 +663,7 @@ static void gen_enum_match(output_t *out, fb_compound_type_t *ct, void *data, in
     unindent(); println(out, "} else {"); indent();
 }
 
-static void gen_scope_match_handler(output_t *out, fb_compound_type_t *unused_ct, void *data, int unused_hint)
+static void gen_scope_match_handler(fb_output_t *out, fb_compound_type_t *unused_ct, void *data, int unused_hint)
 {
     fb_compound_type_t *ct = data;
     fb_scoped_name_t snt;
@@ -678,7 +678,7 @@ static void gen_scope_match_handler(output_t *out, fb_compound_type_t *unused_ct
     println(out, "buf = %s_parse_json_enum(ctx, buf, end, value_type, value, aggregate);", snt.text);
 }
 
-static void gen_scope_match(output_t *out, fb_compound_type_t *ct, void *data, int hint, int n)
+static void gen_scope_match(fb_output_t *out, fb_compound_type_t *ct, void *data, int hint, int n)
 {
     println(out, "buf = flatcc_json_parser_match_scope(ctx, (mark = buf), end, %d);", n);
     println(out, "if (buf != mark) {"); indent();
@@ -686,17 +686,17 @@ static void gen_scope_match(output_t *out, fb_compound_type_t *ct, void *data, i
     unindent(); println(out, "} else {"); indent();
 }
 
-static void gen_field_unmatched(output_t *out)
+static void gen_field_unmatched(fb_output_t *out)
 {
     println(out, "buf = flatcc_json_parser_unmatched_symbol(ctx, buf, end);");
 }
 
-static void gen_enum_unmatched(output_t *out)
+static void gen_enum_unmatched(fb_output_t *out)
 {
     println(out, "return unmatched;");
 }
 
-static void gen_scope_unmatched(output_t *out)
+static void gen_scope_unmatched(fb_output_t *out)
 {
     println(out, "return unmatched;");
 }
@@ -745,8 +745,8 @@ static void gen_scope_unmatched(output_t *out)
 enum trie_type { table_trie, struct_trie, enum_trie, local_scope_trie, global_scope_trie };
 typedef struct trie trie_t;
 
-typedef void gen_match_f(output_t *out, fb_compound_type_t *ct, void *data, int hint, int n);
-typedef void gen_unmatched_f(output_t *out);
+typedef void gen_match_f(fb_output_t *out, fb_compound_type_t *ct, void *data, int hint, int n);
+typedef void gen_unmatched_f(fb_output_t *out);
 
 struct trie {
     dict_entry_t *dict;
@@ -796,7 +796,7 @@ struct trie {
  * is called correctly. This significantly reduces the branching in a
  * case like "Red, Green, Blue".
  */
-static void gen_prefix_trie(output_t *out, trie_t *trie, int a, int b, int pos)
+static void gen_prefix_trie(fb_output_t *out, trie_t *trie, int a, int b, int pos)
 {
     int m, n;
     uint64_t tag = 00, mask = 0;
@@ -837,7 +837,7 @@ static void gen_prefix_trie(output_t *out, trie_t *trie, int a, int b, int pos)
     unindent(); println(out, "} /* \"%.*s\" */", len, name);
 }
 
-static void gen_trie(output_t *out, trie_t *trie, int a, int b, int pos)
+static void gen_trie(fb_output_t *out, trie_t *trie, int a, int b, int pos)
 {
     int x, y, k;
     uint64_t tag = 0, mask = 0;
@@ -1025,7 +1025,7 @@ static void gen_trie(output_t *out, trie_t *trie, int a, int b, int pos)
  */
 static void gen_local_scope_parser(void *context, fb_scope_t *scope)
 {
-    output_t *out = context;
+    fb_output_t *out = context;
     int n = 0;
     trie_t trie;
     fb_symbol_text_t scope_name;
@@ -1074,7 +1074,7 @@ static void gen_local_scope_parser(void *context, fb_scope_t *scope)
  * When a local scope is also parsed, it should be tried before the
  * global scope.
  */
-static int gen_global_scope_parser(output_t *out)
+static int gen_global_scope_parser(fb_output_t *out)
 {
     int n = 0;
     trie_t trie;
@@ -1133,7 +1133,7 @@ static int gen_global_scope_parser(output_t *out)
  * with a binary `or` operation so `pval` must be initialized
  * to 0 before teh first constant in a list.
  */
-static int gen_enum_parser(output_t *out, fb_compound_type_t *ct)
+static int gen_enum_parser(fb_output_t *out, fb_compound_type_t *ct)
 {
     fb_scoped_name_t snt;
     int n = 0;
@@ -1188,7 +1188,7 @@ static int gen_enum_parser(output_t *out, fb_compound_type_t *ct)
  * in. As long as we get input from our own parser we should, however,
  * be reasonable safe as nesting is bounded.
  */
-static int gen_struct_parser(output_t *out, fb_compound_type_t *ct)
+static int gen_struct_parser(fb_output_t *out, fb_compound_type_t *ct)
 {
     fb_scoped_name_t snt;
     int n;
@@ -1247,7 +1247,7 @@ static int gen_struct_parser(output_t *out, fb_compound_type_t *ct)
  * The table end call is omitted in the builder such that callee
  * can do this and get the table reference when and where it is needed.
  */
-static int gen_table_parser(output_t *out, fb_compound_type_t *ct)
+static int gen_table_parser(fb_output_t *out, fb_compound_type_t *ct)
 {
     fb_scoped_name_t snt;
     fb_member_t *member;
@@ -1353,7 +1353,7 @@ static int gen_table_parser(output_t *out, fb_compound_type_t *ct)
  * Like ordinary tables, the table end call is not executed such that
  * callee can get the reference by ending the table.
  */
-static int gen_union_parser(output_t *out, fb_compound_type_t *ct)
+static int gen_union_parser(fb_output_t *out, fb_compound_type_t *ct)
 {
     fb_scoped_name_t snt, snref;
     fb_symbol_t *sym;
@@ -1410,7 +1410,7 @@ static int gen_union_parser(output_t *out, fb_compound_type_t *ct)
 
 static void gen_local_scope_prototype(void *context, fb_scope_t *scope)
 {
-    output_t *out = context;
+    fb_output_t *out = context;
     fb_symbol_text_t scope_name;
 
     fb_copy_scope(scope, scope_name);
@@ -1420,7 +1420,7 @@ static void gen_local_scope_prototype(void *context, fb_scope_t *scope)
     println(out, "int *value_type, uint64_t *value, int *aggregate);");
 }
 
-static int gen_root_table_parser(output_t *out, fb_compound_type_t *ct)
+static int gen_root_table_parser(fb_output_t *out, fb_compound_type_t *ct)
 {
     fb_scoped_name_t snt;
 
@@ -1456,7 +1456,7 @@ static int gen_root_table_parser(output_t *out, fb_compound_type_t *ct)
     return 0;
 }
 
-static int gen_root_struct_parser(output_t *out, fb_compound_type_t *ct)
+static int gen_root_struct_parser(fb_output_t *out, fb_compound_type_t *ct)
 {
     fb_scoped_name_t snt;
 
@@ -1493,7 +1493,7 @@ static int gen_root_struct_parser(output_t *out, fb_compound_type_t *ct)
 }
 
 
-static int gen_root_parser(output_t *out)
+static int gen_root_parser(fb_output_t *out)
 {
     fb_symbol_t *root_type = out->S->root_type.type;
 
@@ -1513,7 +1513,7 @@ static int gen_root_parser(output_t *out)
     return 0;
 }
 
-static int gen_json_parser_prototypes(output_t *out)
+static int gen_json_parser_prototypes(fb_output_t *out)
 {
     fb_symbol_t *sym;
     fb_scoped_name_t snt;
@@ -1580,7 +1580,7 @@ static int gen_json_parser_prototypes(output_t *out)
     return 0;
 }
 
-static int gen_json_parsers(output_t *out)
+static int gen_json_parsers(fb_output_t *out)
 {
     fb_symbol_t *sym;
 
@@ -1607,7 +1607,7 @@ static int gen_json_parsers(output_t *out)
     return 0;
 }
 
-int fb_gen_c_json_parser(output_t *out)
+int fb_gen_c_json_parser(fb_output_t *out)
 {
     gen_json_parser_pretext(out);
     gen_json_parser_prototypes(out);