+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Change Log
 
-## [0.3.5-dev]
-- Better build system support: Add -d option to produce dependency file
-  and allow some test cases to take file arguments.
+## [0.3.5-pre]
+- Refactoring of C code generator to support features below.
+- Add --concat option similar to --stdout, but to a file.
+- Add --depfile and --deptarget options to produce dependency.
+- Allow some test cases to accept arguments over hardcoded path.
 
 ## [0.3.4]